[finance_report_audit] refactor(ledger): move journal write pipeline into ledger.store, dissolve ledger↔accounting cycle (AC12.34.6)
The previous logical fold left a package-level cycle: ledger.ops.post imported
the create/post pipeline from services.accounting, while accounting imported
ledger (Entry) for opening_balance. Runtime-safe via lazy imports, but not a DAG.

Move the journal write pipeline — create_journal_entry / post_journal_entry /
void_journal_entry plus the balance/fx/ownership validators and the
AccountingError/ValidationError types — into ledger/store/posting.py. Now:

- ledger.ops.post depends DOWN on ledger.store.posting (no upward edge);
- services.accounting RE-EXPORTS the pipeline (+ errors) so all ~20 external
  callers and `except ValidationError` sites are unchanged (same class identity);
- ledger has no edge into services.accounting → the cycle is gone.

Module-level cycle check: was `services.accounting ↔ ledger.ops`; now the only
remaining cycle is the benign `models.account ↔ models.journal` TYPE_CHECKING
pair. The backend import graph is a runtime DAG.

Behaviour-preserving: re-export identity verified; 841 accounting/ledger/
investment/fx/reconciliation tests pass. AC12.34.6 added ({tier:PC}).
