+++ apps/backend/src/ledger/store/__init__.py
@@ -1,0 +1,34 @@
+"""Ledger persistence layer (store role).
+
+Owns the journal write pipeline (`create_journal_entry` / `post_journal_entry` /
+`void_journal_entry`) and its validators. `services.accounting` re-exports these
+for its existing callers; `ledger.ops.post_entry` consumes them directly. Moving
+the pipeline here removes the `ledger.ops → services.accounting → ledger` import
+cycle (ledger no longer depends upward on a service).
+"""
+
+from __future__ import annotations
+
+from src.ledger.store.posting import (
+    AccountingError,
+    ValidationError,
+    create_journal_entry,
+    post_journal_entry,
+    validate_fx_rates,
+    validate_journal_balance,
+    validate_journal_posting_invariants,
+    validate_line_account_ownership,
+    void_journal_entry,
+)
+
+__all__ = [
+    "AccountingError",
+    "ValidationError",
+    "create_journal_entry",
+    "post_journal_entry",
+    "validate_fx_rates",
+    "validate_journal_balance",
+    "validate_journal_posting_invariants",
+    "validate_line_account_ownership",
+    "void_journal_entry",
+]
