+++ docs/project/EPIC-012.foundation-libs.md
@@ -538,6 +538,7 @@ runtime `abs(debit-credit) < 0.01` checks.
 | AC12.34.3 | Layer-DAG rule: the model layer imports no service (no upward edge / import cycle); `derive_confidence_tier` lives in the model and the service re-exports it (the old `models.journal → services.confidence_tier` cycle is removed) | `test_AC12_34_3_model_layer_never_imports_a_service`, `test_AC12_34_3_confidence_tier_lives_in_model_layer` | `tests/tooling/test_ledger_module.py` | P1 |
 | AC12.34.4 | Adoption: investment buy/sell/dividend build typed `Entry` and post via `ledger.post_entry` instead of hand-rolling balanced `lines_data` dicts (the legacy `create_journal_entry`/`_post_and_load` path is gone from the service) {tier:PC} | `test_AC12_34_4_investment_postings_use_ledger_post` | `tests/tooling/test_ledger_module.py` | P1 |
 | AC12.34.5 | Every posting path is balance-guaranteed by `Entry`: opening-balance and fx-revaluation construct an `Entry` before persisting (the latter previously wrote raw `JournalLine`s with no balance validation at all), closing the double-entry-integrity gap on the last two paths {tier:PC} | `test_AC12_34_5_remaining_posting_paths_guard_balance_with_entry` | `tests/tooling/test_ledger_module.py` | P1 |
+| AC12.34.6 | The journal write pipeline (`create_journal_entry`/`post_journal_entry`/`void_journal_entry` + validators) lives in `ledger.store`; `ledger.ops` depends down on it instead of up on `services.accounting`, dissolving the `ledger ↔ services.accounting` import cycle. `services.accounting` re-exports the pipeline so external callers and `except ValidationError` are unchanged {tier:PC} | `test_AC12_34_6_ledger_owns_posting_pipeline_no_upward_edge` | `tests/tooling/test_ledger_module.py` | P1 |
 
 ---
